refactor(agent): extract catalog tool contract
