+++ AGENTS.md
@@ -44,8 +44,10 @@ Top-level orchestration is via `docker-compose.yaml`; optional local NIM model c
 
 ## 3) Source Map (Where to Change What)
 
-- Serving agent orchestration, registered tools, and capability-derived
-  model-visible search schema: `chain_server/src/deepagents_runtime.py`
+- Serving agent orchestration and registered tool wrappers:
+  `chain_server/src/deepagents_runtime.py`
+- Pure capability-derived catalog tool schema, structural taxonomy mapping,
+  and canonical scope identity: `chain_server/src/catalog_tool_contract.py`
 - Reusable model-visible catalog search rules: `chain_server/src/catalog_scope.py`
 - Shopper-skill registry, frontmatter validation, and immutable tool policy: `chain_server/src/tool_policy.py`
 - Per-turn skill activation, model-visible tool binding, and dispatch grant gate: `chain_server/src/skill_activation.py`
