+++ chain_server/src/deepagents_runtime.py
@@ -26,7 +26,6 @@ from pydantic import (
     Field,
     ValidationError,
     create_model,
-    field_validator,
     model_validator,
 )
 from pydantic_core import PydanticCustomError
@@ -35,7 +34,6 @@ import requests
 from .agenttypes import Cart, ShopperContext, State
 from .catalog_capabilities import (
     CatalogCapabilitiesClient,
-    effective_filter_capabilities,
     format_catalog_capabilities_for_prompt,
 )
 from .catalog_execution import execute_catalog_search
@@ -45,6 +43,15 @@ from .catalog_request import (
     build_catalog_search_plan,
 )
 from .catalog_scope import CATALOG_SEARCH_RULES
+from .catalog_tool_contract import (
+    CatalogTaxonomyToolInput as CatalogTaxonomyToolInput,
+    SearchCatalogToolArguments,
+    SearchCatalogToolInput as SearchCatalogToolInput,
+    build_search_catalog_tool_input_model as _search_catalog_tool_input_model,
+    catalog_search_scope as _catalog_search_scope,
+    normalize_catalog_scope_value as _normalized_scope_value,
+    taxonomy_hard_constraints as _taxonomy_hard_constraints,
+)
 from .commerce_tools import (
     add_cart_item,
     check_active_promotions,
@@ -378,38 +385,6 @@ _INTERNAL_SHOPPER_REPLACEMENTS = (
 )
 
 
-class CatalogTaxonomyToolInput(BaseModel):
-    """Catalog-derived taxonomy roles used by the agent search tool."""
-
-    model_config = ConfigDict(extra="forbid")
-
-    category: list[str] = Field(
-        ...,
-        description=(
-            "Exact advertised category values required by the shopper. Use an "
-            "empty list only when subcategory supplies the text-search scope or "
-            "the search is image-only."
-        ),
-    )
-    subcategory: list[str] = Field(
-        ...,
-        description=(
-            "Exact advertised subcategory values required by the shopper. Use an "
-            "empty list when category supplies the text-search scope or the search "
-            "is image-only."
-        ),
-    )
-
-    @field_validator("category", "subcategory", mode="before")
-    @classmethod
-    def deduplicate_values(cls, value: Any) -> Any:
-        """Normalize repeated taxonomy values before cardinality validation."""
-
-        if isinstance(value, str):
-            return [value]
-        return list(dict.fromkeys(value)) if isinstance(value, list) else value
-
-
 def _generic_shopper_guidance(requested_product_type: str | None) -> str:
     """Return safe guidance after an inferred attribute is removed."""
 
@@ -519,369 +494,6 @@ def _products_with_subcategory_coverage(
     return [products[index] for index in sorted(selected_indexes)]
 
 
-class SearchCatalogToolArguments(BaseModel):
-    """Stable agent-facing catalog-search arguments."""
-
-    model_config = ConfigDict(extra="forbid")
-
-    semantic_query: str = Field(
-        ...,
-        description=(
-            "One soft or descriptive product search string for semantic ranking. "
-            "Product type may be repeated for relevance, but taxonomy and other "
-            "must-haves are enforced only by the structured fields below. Use an "
-            "empty string only for an image-only search."
-        ),
-    )
-    shopper_guidance: str = Field(
-        ...,
-        max_length=400,
-        description=(
-            "One concise, product-agnostic shopper-facing sentence written under "
-            "the active skill before search results are known. Connect this "
-            "product role to the shopper's stated goal or direct antecedent. Do "
-            "not name unselected or unavailable product types, name or describe "
-            "candidate products, assert product attributes, or mention tools, "
-            "schemas, filters, evidence, or identifiers. Use an empty string only "
-            "for image-only search."
-        ),
-    )
-    requested_product_type: str | None = Field(
-        ...,
-        description=(
-            "Shortest product noun or umbrella phrase for this focused role, "
-            "resolved from the shopper's current turn or direct antecedent. "
-            "Exclude color, material, fit, occasion, weather, and style modifiers: "
-            "'formal tops' and 'relaxed-fit tops' both use 'tops'. For a genuinely "
-            "open role selected by the agent, use the chosen advertised role noun. "
-            "If this type is not separately advertised and you select one faithful "
-            "advertised parent category, keep this shopper-named type unchanged. "
-            "This is provenance, not catalog taxonomy or a ranking query. Use null "
-            "only for image-only search."
-        ),
-    )
-    taxonomy: CatalogTaxonomyToolInput = Field(
-        ...,
-        description=(
-            "Required catalog-derived taxonomy selection. Allowed category and "
-            "subcategory values come from the active catalog capabilities. Every "
-            "selected value must be the requested product type or a child of an "
-            "umbrella the shopper actually named. For a genuinely open request, "
-            "select one advertised subcategory as the focused role. Never select "
-            "a parent or sibling as a substitute for an advertised type. If a "
-            "shopper-named type is not separately advertised but one advertised "
-            "category is its faithful broader parent, select only that category "
-            "and leave subcategory empty; results remain alternatives under their "
-            "actual catalog types. For example, skirts may satisfy bottoms; "
-            "dresses may not."
-        ),
-    )
-    required_constraints: dict[str, Any] = Field(
-        ...,
-        description=(
-            "Every non-taxonomy must-have stated for the target products in the "
-            "current shopper turn, as a structured field and value. Facts about "
-            "an antecedent or anchor guide semantic styling judgment; do not copy "
-            "them onto a complementary product unless the shopper explicitly asks "
-            "for the same value. "
-            "Use exact hard-filter properties and advertised enum values from "
-            "current Catalog capabilities. Put a directly stated requirement "
-            "those capabilities do not advertise in unadvertised_requirements. "
-            "Season, weather, occasion, "
-            "and subjective style/vibe "
-            "context remain semantic direction unless the shopper directly "
-            "requires an objective product attribute. A defining material is a "
-            "must-have: for 'Any denim skirts available?', put 'denim' in "
-            "unadvertised_requirements when composition is not a hard filter. "
-            "Product types never belong in unadvertised_requirements."
-        ),
-    )
-    scope_complete: bool = Field(
-        ...,
-        description=(
-            "True only when this search plus existing turn evidence is enough to "
-            "answer the shopper's complete current request. For a recommendation-"
-            "only one-role request this is true. Set false when an explicitly "
-            "requested product role, product-detail verification, availability "
-            "check, or cart action still must run after this search. Do not set "
-            "false merely to search alternatives or adjacent types, or because a "
-            "broader multi-turn outfit project remains unfinished. 'Start with a "
-            "beige top' and 'What bottoms go with that?' are each complete after "
-            "their own one-role search."
-        ),
-    )
-    search_mode: str | None = Field(
-        default=None,
-        description="Optional search mode from Catalog capabilities.",
-    )
-
-
-class SearchCatalogToolInput(SearchCatalogToolArguments):
-    """Catalog search request validated only against structural capabilities."""
-
-    @model_validator(mode="after")
-    def text_search_has_taxonomy_scope(self) -> "SearchCatalogToolInput":
-        if len(set(self.taxonomy.category)) > 1:
-            raise ValueError("catalog search accepts at most one category")
-        has_taxonomy = bool(self.taxonomy.category or self.taxonomy.subcategory)
-        has_query = bool(self.semantic_query.strip())
-        has_shopper_guidance = bool(self.shopper_guidance.strip())
-        requested_product_type = (
-            self.requested_product_type.strip()
-            if isinstance(self.requested_product_type, str)
-            else ""
-        )
-        image_only = not has_query and not has_taxonomy and not requested_product_type
-        if image_only:
-            if has_shopper_guidance:
-                raise ValueError(
-                    "image-only search requires empty shopper_guidance"
-                )
-            constraints = (
-                self.required_constraints.model_dump(exclude_none=True)
-                if isinstance(self.required_constraints, BaseModel)
-                else self.required_constraints
-            )
-            if any(value not in (None, "", [], {}) for value in constraints.values()):
-                raise ValueError(
-                    "image-only search requires empty required_constraints"
-                )
-            return self
-        if not requested_product_type:
-            raise ValueError("text catalog search requires requested_product_type")
-        if not has_shopper_guidance:
-            raise ValueError("catalog retrieval requires non-empty shopper_guidance")
-        if not has_taxonomy:
-            raise ValueError(
-                "text catalog search requires an advertised category or subcategory"
-            )
-        if not has_query:
-            raise ValueError("text catalog search requires a semantic query")
-        return self
-
-
-class _CatalogNumberConstraint(BaseModel):
-    """Inclusive numeric range accepted by catalog hard filters."""
-
-    model_config = ConfigDict(extra="forbid")
-
-    min: float | None = None
-    max: float | None = None
-
-    @model_validator(mode="after")
-    def has_a_bound(self) -> "_CatalogNumberConstraint":
-        if self.min is None and self.max is None:
-            raise ValueError("numeric constraint requires min and/or max")
-        if self.min is not None and self.max is not None and self.min > self.max:
-            raise ValueError("numeric constraint min cannot exceed max")
-        return self
-
-
-def _search_catalog_tool_input_model(
-    capabilities: CatalogCapabilities,
-    *,
-    validate_scope: bool = True,
-) -> type[SearchCatalogToolArguments]:
-    """Create one tool schema whose enums come from the active catalog."""
-
-    taxonomy = capabilities.taxonomy
-    advertised_search_modes = tuple(dict.fromkeys(capabilities.retrieval_modes))
-    search_mode_type = (
-        Literal.__getitem__(advertised_search_modes)
-        if advertised_search_modes
-        else str
-    )
-    category_values = (
-        sorted(taxonomy.categories, key=str.casefold)
-        if taxonomy.category_field
-        else []
-    )
-    subcategory_values = (
-        sorted(
-            {
-                name
-                for category in taxonomy.categories.values()
-                for name in category.subcategories
-            },
-            key=str.casefold,
-        )
-        if taxonomy.subcategory_field
-        else []
-    )
-
-    category_type, category_field = _taxonomy_list_field(
-        category_values,
-        role="category",
-        advertised_field=taxonomy.category_field,
-    )
-    subcategory_type, subcategory_field = _taxonomy_list_field(
-        subcategory_values,
-        role="subcategory",
-        advertised_field=taxonomy.subcategory_field,
-    )
-    taxonomy_model = create_model(
-        "CatalogTaxonomySelection",
-        __base__=CatalogTaxonomyToolInput,
-        category=(category_type, category_field),
-        subcategory=(subcategory_type, subcategory_field),
-    )
-    required_constraints_model = _required_constraints_input_model(
-        capabilities,
-    )
-    return create_model(
-        "CatalogSearchInput" if validate_scope else "CatalogSearchToolArguments",
-        __base__=(
-            SearchCatalogToolInput
-            if validate_scope
-            else SearchCatalogToolArguments
-        ),
-        taxonomy=(
-            taxonomy_model,
-            Field(
-                ...,
-                description=(
-                    "Required taxonomy selection generated from the active catalog. "
-                    "Use exact enum values. A text search requires at least one "
-                    "category or subcategory; both arrays may be empty only for "
-                    "an image-only search."
-                    " Every value must be a kind of the requested scope: skirts "
-                    "may satisfy bottoms; dresses may not. For a broad request "
-                    "that names no product type, choose one exact advertised "
-                    "subcategory as the focused starting role. If a shopper-named "
-                    "type is not separately advertised but one faithful broader "
-                    "advertised parent category exists, select only that category "
-                    "and leave subcategory empty."
-                ),
-            ),
-        ),
-        required_constraints=(
-            required_constraints_model,
-            Field(
-                ...,
-                description=(
-                    "Catalog hard filters and any defining requirement the active "
-                    "catalog cannot enforce. Apply constraints only when the "
-                    "current turn states them for the target products; an anchor's "
-                    "attributes belong in semantic styling context unless the "
-                    "shopper explicitly requests the same value. Use only the "
-                    "advertised properties "
-                    "in this object; put unsupported must-haves in "
-                    "unadvertised_requirements instead of weakening them. For "
-                    "'Do you have water-resistant bags?', use "
-                    "{'unadvertised_requirements': ['water resistance']}. Do not "
-                    "put broad season, weather, occasion, or subjective style/vibe "
-                    "context here unless the shopper directly requires a product "
-                    "attribute. 'Rainy day outfit' does not require water "
-                    "resistance; 'water-resistant bags' does. Recommendation "
-                    "adjectives such as comfortable, relaxed, soft, breathable, "
-                    "lightweight, casual, dressy, bold, bright, vibrant, or "
-                    "sporty are always semantic ranking preferences, not "
-                    "objective hard filters. Before calling the tool, compare "
-                    "every target-product modifier with this advertised schema "
-                    "and include every exact matching filter value. A product type "
-                    "never belongs in unadvertised_requirements. Use an empty "
-                    "object for image-only search."
-                ),
-            ),
-        ),
-        search_mode=(
-            search_mode_type | None,
-            Field(
-                default=None,
-                description="Optional search mode advertised by the active catalog.",
-            ),
-        ),
-    )
-
-
-def _taxonomy_list_field(
-    values: list[str],
-    *,
-    role: str,
-    advertised_field: str | None,
-) -> tuple[Any, Any]:
-    field_name = advertised_field or "not advertised"
-    description = (
-        f"Exact {role} values advertised through catalog field '{field_name}'. "
-        "Use an empty list when the other taxonomy role supplies the text scope or "
-        "the search is image-only."
-    )
-    if role == "category":
-        description += " Select at most one category per catalog search."
-    if not values:
-        return list[str], Field(..., max_length=0, description=description)
-
-    literal_type = Literal.__getitem__(tuple(values))
-    max_length = 1 if role == "category" else None
-    return list[literal_type], Field(
-        ...,
-        max_length=max_length,
-        description=description,
-    )
-
-
-def _required_constraints_input_model(
-    capabilities: CatalogCapabilities,
-) -> type[BaseModel]:
-    """Create a hard-constraint schema from advertised catalog filters."""
-
-    taxonomy_fields = {
-        field_name
-        for field_name in (
-            capabilities.taxonomy.category_field,
-            capabilities.taxonomy.subcategory_field,
-        )
-        if field_name
-    }
-    fields: dict[str, tuple[Any, Any]] = {}
-    for name, capability in sorted(effective_filter_capabilities(capabilities).items()):
-        if name in taxonomy_fields:
-            continue
-        if capability.type in {"enum", "enum_list"} and capability.values:
-            value_type = Literal.__getitem__(tuple(capability.values))
-            field_type = value_type | list[value_type] | None
-        elif capability.type == "number":
-            field_type = _CatalogNumberConstraint | None
-        else:
-            field_type = str | list[str] | None
-        fields[name] = (
-            field_type,
-            Field(
-                default=None,
-                description=f"Advertised hard filter '{name}'.",
-            ),
-        )
-    fields["unadvertised_requirements"] = (
-        list[str],
-        Field(
-            default_factory=list,
-            description=(
-                "Only objective product requirements directly stated in the "
-                "current shopper turn and absent from this schema. Never infer a "
-                "requirement from season, weather, occasion, or subjective "
-                "style/vibe context: 'rainy day outfit' produces an empty list. "
-                "Use ['water resistance'] for 'water-resistant bags' because that "
-                "turn directly states the product requirement. Never omit or "
-                "soften a directly stated objective requirement. Subjective "
-                "recommendation adjectives such as comfortable, relaxed, soft, "
-                "breathable, lightweight, casual, dressy, bold, bright, vibrant, "
-                "or sporty always stay semantic; never put them in this field. "
-                "Before calling the tool, include every exact advertised filter "
-                "value that modifies the target product; do not leave this object "
-                "empty when one applies. "
-                "In an 'A or B' request, do not put the "
-                "supported advertised branch here. A product type never belongs "
-                "here."
-            ),
-        ),
-    )
-    return create_model(
-        "CatalogRequiredConstraints",
-        __config__=ConfigDict(extra="forbid"),
-        **fields,
-    )
-
-
 class _ShopperSkillActivationInput(BaseModel):
     """Shared composition rules for dynamic shopper-skill activation."""
 
@@ -935,108 +547,6 @@ def _skill_activation_input_model(
             ),
         ),
     )
-
-
-def _taxonomy_hard_constraints(
-    selection: BaseModel | dict[str, Any],
-    capabilities: CatalogCapabilities,
-) -> tuple[dict[str, list[str]], list[str]]:
-    """Map generic taxonomy roles to catalog-owned hard-filter field names."""
-
-    raw = selection.model_dump() if isinstance(selection, BaseModel) else selection
-    categories = sorted(set(raw.get("category", [])), key=str.casefold)
-    subcategories = sorted(set(raw.get("subcategory", [])), key=str.casefold)
-    taxonomy = capabilities.taxonomy
-    issues: list[str] = []
-
-    for category in categories:
-        if category not in taxonomy.categories:
-            issues.append(f"category '{category}' is not advertised")
-
-    owners: dict[str, list[str]] = {
-        subcategory: sorted(
-            [
-                category_name
-                for category_name, category in taxonomy.categories.items()
-                if subcategory in category.subcategories
-            ],
-            key=str.casefold,
-        )
-        for subcategory in subcategories
-    }
-    for subcategory, subcategory_owners in owners.items():
-        if not subcategory_owners:
-            issues.append(f"subcategory '{subcategory}' is not advertised")
-        elif categories and not set(categories).intersection(subcategory_owners):
-            issues.append(
-                f"subcategory '{subcategory}' is not available in selected categories"
-            )
-
-    if categories and subcategories:
-        for category in categories:
-            advertised = taxonomy.categories.get(category)
-            if advertised is None:
-                continue
-            if not set(subcategories).intersection(advertised.subcategories):
-                issues.append(
-                    f"category '{category}' has no selected subcategory"
-                )
-
-    effective_categories = categories
-    if subcategories and not categories:
-        inferred_categories = sorted(
-            {
-                owner
-                for subcategory_owners in owners.values()
-                for owner in subcategory_owners
-            },
-            key=str.casefold,
-        )
-        if len(inferred_categories) > 1:
-            issues.append(
-                "subcategory selection has multiple owning categories; select "
-                "exactly one advertised category"
-            )
-            effective_categories = []
-        else:
-            effective_categories = inferred_categories
-
-    constraints: dict[str, list[str]] = {}
-    if effective_categories:
-        if taxonomy.category_field:
-            constraints[taxonomy.category_field] = effective_categories
-        else:
-            issues.append("the catalog does not advertise a category filter field")
-    if subcategories:
-        if taxonomy.subcategory_field:
-            constraints[taxonomy.subcategory_field] = subcategories
-        else:
-            issues.append("the catalog does not advertise a subcategory filter field")
-    return constraints, issues
-
-
-def _catalog_search_scope(
-    taxonomy: dict[str, list[str]],
-    required_constraints: dict[str, Any],
-) -> dict[str, Any]:
-    """Return the normalized hard-filter identity for one catalog search."""
-
-    return {
-        "taxonomy": _normalized_scope_value(taxonomy),
-        "required_constraints": _normalized_scope_value(required_constraints),
-    }
-
-
-def _normalized_scope_value(value: Any) -> Any:
-    if isinstance(value, dict):
-        return {
-            key: _normalized_scope_value(item)
-            for key, item in sorted(value.items())
-        }
-    if isinstance(value, list):
-        normalized = [_normalized_scope_value(item) for item in value]
-        return sorted(normalized, key=lambda item: json.dumps(item, sort_keys=True))
-    return value
 
 
 class AddCartItemsToolItemInput(BaseModel):
