+++ tests/unit/chain_server/test_catalog_scope.py
@@ -5,9 +5,24 @@
 
 
 def test_model_catalog_search_has_no_semantic_relation_label() -> None:
-    from chain_server.src.deepagents_runtime import SearchCatalogToolArguments
+    from chain_server.src.catalog_tool_contract import SearchCatalogToolArguments
 
     assert "taxonomy_status" not in SearchCatalogToolArguments.model_fields
+
+
+def test_catalog_search_scope_identity_is_order_independent() -> None:
+    from chain_server.src.catalog_tool_contract import catalog_search_scope
+
+    first = catalog_search_scope(
+        {"subcategory": ["heels", "flats"], "category": ["footwear"]},
+        {"price": {"max": 100, "min": 25}, "primary_color": ["navy", "black"]},
+    )
+    second = catalog_search_scope(
+        {"category": ["footwear"], "subcategory": ["flats", "heels"]},
+        {"primary_color": ["black", "navy"], "price": {"min": 25, "max": 100}},
+    )
+
+    assert first == second
 
 
 def test_catalog_search_rules_allow_model_selected_parent_category() -> None:
